[ssr][pictures] refactored to fresh common

+++ archivy/render/picture.py
@@ -1,4 +1,4 @@
-from archivy.render.common import default_handlers, handler_info, handler_engine
+from archivy.render.common import default_handlers, handler_info, handler_engine, sync_files
 import os
 import shutil
 
@@ -16,15 +16,28 @@ def render_picture(
     if src == "":
         raise ValueError("picture supports input from files only!")
 
-    if os.path.exists(d_path):
-        os.unlink(d_path)
-    if not os.path.exists(src):
-        return False, [f"Source image '{src}' is not found!"]
-    shutil.copy2(src, d_path)
-
     result = True, None
 
-    return result
+    if os.path.isdir(src):
+        result[0] = False
+        result[1].append([f"src '{src}' should point to a file!"])
+
+    if not os.path.exists(src):
+        result[0] = False
+        result[1].append([f"Source picture '{src}' is not found!"])
+
+    if not result[0]:
+        if os.path.exists(d_path):
+            if os.path.isfile(d_path):
+                os.unlink(d_path)
+            else:
+                shutil.rmtree(d_path)
+        return result
+
+    if sync_files(src, d_path, allow_delete=True):
+        return result
+    else:
+        return False, [f"Failed to sync '{src}' to '{d_path}'"]
 
 
 default_handlers.register_handler(
